error solving
